relocate DRep to primitive lib
